refactor(search): 重构搜索模块
重构搜索模块结构，将cogs移动到search目录下
